requirements.txt
blinker==1.9.0
bootstrap4==0.1.0
click==8.3.1
colorama==0.4.6
dominate==2.9.1
Flask==3.1.2
Flask-Admin==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Modals==0.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
